src/requirements.txt
jinja2
python-dateutil
pytz
parameterized
freezegun
